requirements.txt
boto3==1.34.156
requests~=2.32.3

pyspark~=3.5.1
pandas~=2.2.2
python-dateutil~=2.9.0.post0